:recycle: Fixes most style-related issues
