+++ lib/components/carousel/bullet.styles.ts
@@ -1,16 +1,17 @@
 import { StyleSheet } from 'react-native'
+import type { ThemeFunction } from '@styles/types'
 
 const bulletSize = 6
 const bulletSizeBig = bulletSize + 2
 
-const styles = (theme) => ({
+const styles: ThemeFunction = ({ colors }) => ({
   bullet: {
     height: bulletSize,
     width: bulletSize,
     borderRadius: bulletSize / 2,
     marginHorizontal: 5,
     opacity: 0.5,
-    backgroundColor: theme.colors.backgroundAlt,
+    backgroundColor: colors.backgroundAlt,
   },
   bulletActive: {
     opacity: 0.6,
@@ -22,7 +23,7 @@ const styles = (theme) => ({
     height: bulletSizeBig,
     width: bulletSizeBig,
     borderRadius: bulletSizeBig / 2,
-    borderColor: theme.colors.textAlt,
+    borderColor: colors.textAlt,
     borderStyle: 'solid',
     borderWidth: StyleSheet.hairlineWidth,
   },
