+++ lib/components/welcome/index.tsx
@@ -3,7 +3,7 @@ import { useTheme } from '@hooks'
 import Header from '@components/welcome/header'
 import Intro from '@components/welcome/intro'
 import Actions from '@components/welcome/actions'
-import { STRUCTURE_TAB_NAV_HEIGHT } from '@utils/constants'
+import welcomeComponentStyle from './index.styles'
 
 const Welcome = () => {
   const {
@@ -12,26 +12,7 @@ const Welcome = () => {
     theme: { colors },
   } = useTheme()
 
-  const styles = createStyles((theme) => ({
-    container: {
-      flex: 1,
-      backgroundColor: theme.colors.neutralLight,
-    },
-    header: {
-      flex: 1,
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    intro: {
-      flex: 6,
-      alignItems: 'center',
-      justifyContent: 'center',
-    },
-    actions: {
-      alignItems: 'center',
-      marginBottom: STRUCTURE_TAB_NAV_HEIGHT / 2,
-    },
-  }))
+  const styles = createStyles(welcomeComponentStyle)
 
   return (
     <SafeAreaView style={styles.container}>
