+++ lib/screens/messages/private-messages-screen-header.tsx
@@ -1,54 +1,19 @@
+import type { FC } from 'react'
 import { Text, View } from 'react-native'
 import { useTheme } from '@hooks'
 import ScreenSimpleHeaderTemplate from '@templates/screen-simple-header'
 import { UIAvatar } from '@ui-elements'
 import { ChatOptions } from '@screens/messages/inner-components'
 import { littenToHeaderTitle, shortenName } from '@utils/functions'
-import {
-  PLACEHOLDER_USER_DISPLAY_NAME,
-  USER_AVATAR_SIZE_MINI,
-} from '@utils/constants'
+import { PLACEHOLDER_USER_DISPLAY_NAME } from '@utils/constants'
+import messagePrivateHeaderComponentStyles from '@screens/messages/private-messages-screen-header.styles'
+import { ChatOptionsProps } from './inner-components/chat-options'
 
-const MessagePrivateHeader = (props) => {
+const MessagePrivateHeader: FC<ChatOptionsProps> = (props) => {
   const { litten, recipient } = props
   const { createStyles } = useTheme()
 
-  const styles = createStyles((theme, typography) => ({
-    headerContainer: {
-      flex: 1,
-      flexDirection: 'row',
-      alignItems: 'flex-end',
-      justifyContent: 'flex-start',
-      top: USER_AVATAR_SIZE_MINI / 7,
-    },
-    titleContainer: {
-      flex: 1,
-      alignItems: 'flex-start',
-    },
-    messagesContainer: {
-      flex: 1,
-      backgroundColor: theme.colors.neutralLight,
-    },
-    avatarContainer: {
-      marginRight: USER_AVATAR_SIZE_MINI / 3,
-    },
-    title: {
-      textAlign: 'left',
-      fontSize: typography.fontSize.base,
-      fontWeight: typography.fontWeight.bolder,
-      color: theme.colors.textAlt,
-    },
-    subtitle: {
-      textAlign: 'left',
-      fontSize: typography.fontSize.base,
-      fontWeight: typography.fontWeight.light,
-      color: theme.colors.textAlt,
-    },
-    conversationOptions: {
-      alignItems: 'center',
-      justifyContent: 'flex-end',
-    },
-  }))
+  const styles = createStyles(messagePrivateHeaderComponentStyles)
 
   const subtitle = littenToHeaderTitle(litten)
