+++ lib/templates/screen-simple-header.tsx
@@ -2,10 +2,7 @@ import { Platform, Text, View } from 'react-native'
 import type { ViewProps } from 'react-native'
 import { useTheme } from '@hooks'
 import GoBack from '@components/go-back'
-import {
-  STRUCTURE_TEMPLATE_SCREEN_HEADER_PADDING_BOTTOM,
-  STRUCTURE_TEMPLATE_SCREEN_HEADER_WIDTH,
-} from '@utils/constants'
+import screenSimpleHeaderTemplateStyles from '@templates/screen-simple-header.styles'
 
 export type ScreenSimpleHeaderTemplateProps = {
   children?: ViewProps['children']
@@ -20,29 +17,7 @@ const ScreenSimpleHeaderTemplate = ({
 }: ScreenSimpleHeaderTemplateProps) => {
   const { createStyles } = useTheme()
 
-  const styles = createStyles((theme, typography) => ({
-    headerContainer: {
-      height: '100%',
-      width: STRUCTURE_TEMPLATE_SCREEN_HEADER_WIDTH,
-      flexDirection: 'row',
-      alignItems: 'flex-end',
-      justifyContent: 'center',
-      paddingBottom: STRUCTURE_TEMPLATE_SCREEN_HEADER_PADDING_BOTTOM,
-    },
-    headerContentContainer: {
-      flex: 1,
-      flexGrow: 16,
-    },
-    headerText: {
-      fontSize: typography.fontSize.xxlarge,
-      fontWeight: typography.fontWeight.bolder,
-      textAlign: 'left',
-      color: theme.colors.textAlt,
-    },
-    headerTextCentered: {
-      textAlign: 'center',
-    },
-  }))
+  const styles = createStyles(screenSimpleHeaderTemplateStyles)
 
   const notAndroid = Platform.OS !== 'android'
   const childrenIsText = typeof children === 'string'
