+++ lib/ui-elements/setting/index.tsx
@@ -1,8 +1,8 @@
-import { Pressable, StyleSheet, Text, View } from 'react-native'
+import { Pressable, Text, View } from 'react-native'
 import type { PressableProps, ViewProps } from 'react-native'
 import { useTheme } from '@hooks'
 import UIText from '@ui-elements/text'
-import { UI_ELEMENT_BORDER_MARGIN } from '@utils/constants'
+import settingComponentStyles from '@ui-elements/setting/index.styles'
 
 type UISettingProps = {
   description?: string
@@ -19,29 +19,7 @@ const UISetting = ({
 }: UISettingProps) => {
   const { createStyles } = useTheme()
 
-  const styles = createStyles((theme, typography) => ({
-    uiSettingContainer: {
-      flexDirection: 'row',
-      alignItems: 'center',
-      justifyContent: 'space-between',
-      borderBottomWidth: StyleSheet.hairlineWidth,
-      borderBottomColor: theme.colors.neutral,
-      paddingTop: 12,
-      paddingBottom: 12,
-      paddingLeft: 8,
-      paddingRight: 8,
-      marginTop: 12,
-    },
-    uiSettingText: {
-      flex: 1,
-      color: theme.colors.text,
-      fontSize: typography.fontSize.large,
-      fontWeight: typography.fontWeight.light,
-    },
-    uiSettingDescription: {
-      marginTop: UI_ELEMENT_BORDER_MARGIN,
-    },
-  }))
+  const styles = createStyles(settingComponentStyles)
 
   const content = (
     <>
